Split test file into per function files
